requirements.txt
dateparser==1.0.0
hdbcli==2.8.20
ktrain==0.29.2
numpy==1.19.5
pandas==1.2.3
python-dotenv==0.18.0
py2neo==2021.1.1
scikit-learn==0.23.2
SQLAlchemy==1.4.17
sqlalchemy-hana==0.5.0
striprtf==0.0.13
tensorflow==2.5.0rc0
tqdm==4.49.0
